server/requirements.txt
anyio==3.6.2
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.92.0
h11==0.14.0
idna==3.4
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.26.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
